server/requirements.txt
asgiref==3.11.0
boto3==1.42.30
botocore==1.42.30
brotli==1.2.0
cffi==2.0.0
cssselect2==0.8.0
Django==5.0.7
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
fonttools==4.61.1
gunicorn==23.0.0
jmespath==1.0.1
packaging==25.0
pillow==12.1.0
pycparser==2.23
pydyf==0.12.1
PyJWT==2.10.1
pyphen==0.17.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
s3transfer==0.16.0
six==1.17.0
sqlparse==0.5.5
tinycss2==1.5.1
tinyhtml5==2.0.0
urllib3==2.6.3
weasyprint==67.0
webencodings==0.5.1
zopfli==0.4.0
